requirements.txt
altgraph==0.17.4
colorama==0.4.6
comtypes==1.4.8
docx2pdf==0.1.8
lxml==5.3.0
packaging==24.2
pefile==2023.2.7
pyasn1==0.6.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
python-docx==1.1.2
pywin32==306
rsa==4.9
setuptools==75.6.0
tqdm==4.67.1
typing_extensions==4.12.2
